requirements.txt
pyrogram~=2.0.106
TgCrypto~=1.2.5
py-tgcalls~=2.2.8
yt-dlp>=2026.3.3
aiohttp~=3.11.0
aiosqlite~=0.21.0
pillow~=11.0.0
psutil~=7.2.0
py-yt-search~=0.5.5
python-dotenv~=1.2.1
